fixed, away msg retrieve from gaim was broken
git-svn-id: http://miranda.googlecode.com/svn/trunk@2227 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/IcqOscarJ/fam_04message.c
@@ -2076,7 +2076,7 @@ static void handleRecvServMsgError(unsigned char *buf, WORD wLen, WORD wFlags, D
 
       packet.wLen = 13 + getUINLen(dwUin);
       write_flap(&packet, ICQ_DATA_CHAN);
-      packFNACHeader(&packet, ICQ_LOCATION_FAMILY, ICQ_LOCATION_REQ_USER_INFO, 0, dwSequence);
+      packFNACHeader(&packet, ICQ_LOCATION_FAMILY, ICQ_LOCATION_REQ_USER_INFO, 0, (WORD)dwSequence);
       packWord(&packet, 0x03);
       packUIN(&packet, dwUin);
